Containerization/requirements.txt
streamlit
lancedb
sentence-transformers
ctransformers
langchain
langchain_community
pypdf
